refactor: structure abci app in a way it can be parsed
