+++ packages/valory/skills/decision_maker_abci/behaviours/decision_maker.py
@@ -30,7 +30,7 @@ from packages.valory.skills.decision_maker_abci.behaviours.base import (
     DecisionMakerBaseBehaviour,
 )
 from packages.valory.skills.decision_maker_abci.payloads import DecisionMakerPayload
-from packages.valory.skills.decision_maker_abci.rounds.decision_maker import (
+from packages.valory.skills.decision_maker_abci.states.decision_maker import (
     DecisionMakerRound,
 )
 from packages.valory.skills.decision_maker_abci.tasks import (
